backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
pymysql
python-dotenv
